Revert to the original parameter stringify method
The way the params were stringified in the original repo is actually preferable, as using your way the created signatures are invalid when your params have spaces or other non-url safe characters in them. I've verified this on the InboundShipments Api methods where I was pushing SKUs containing spaces and forward slashes. With the method like this it works, and with the modified version you either end up double-encoding or you get an invalid signature error and the call fails